Use a proper temp file.
Standalone compiled Matlab programs have a default path wherever they are installed. For most OSes, writing to the program installation directory (\Program Files for Windows; /Applications for OSX) is a privileged operation. For normal users this will fail. Most OSes provide a directory for temporary files generated by user programs, as well as an API to find it. Matlab provides the cross-platform tempdir function for this purpose. This commit allow plot_google_map to be used in in standalone compiled Matlab programs.